data_analysis/event_requirements.txt
mysql-connector-python==9.0.0
numpy==2.1.1
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tzdata==2024.1
